refactor: replace Redis user storage with context-based user holder
- Removed `dto.LoginForm` and `dto.User` in favor of `api.SimpleUser`
- Introduced `user_holder` package to manage user context propagation
- Updated `GetMe` service method to retrieve user from context instead of Redis
- Refactored `Auth` middleware to inject user data into context using `user_holder.WithUser`
- Simplified Redis key handling and TTL management in constants
- Adjusted `SimpleUser` ID type from `int64` to `uint` for consistency

+++ api/user.go
@@ -22,7 +22,7 @@ type (
 
 type (
 	SimpleUser struct {
-		Id       int64  `json:"id"`
+		Id       uint   `json:"id"`
 		NickName string `json:"nickname"`
 		Icon     string `json:"icon"`
 	}
